refactor: skeleton
