+++ src/application/core/exceptions/token.py
