+++ src/application/core/middlewares/authentication_external.py
@@ -15,7 +15,7 @@ from starlette.requests import HTTPConnection
 
 from application.core.exceptions import ExternalServiceException, TokenException
 from application.core.external_service import AuthClient
-from application.domain.auth.exceptions import TokenDecodeException
+from application.core.exceptions.token import TokenDecodeException
 
 from ..exceptions.middleware import NoAuthenticationException
