--- src/application/domain/domain_structure/exceptions.py
@@ -1,5 +1,0 @@
-from application.core.exceptions.base import CustomException
-
-
-class YourBaseException(CustomException):
-    pass
