+++ src/application/domain/domain_structure/repository.py
@@ -1,14 +1,14 @@
 from dependency_injector.wiring import Provide
 from sqlalchemy.ext.asyncio import async_scoped_session
 
-from application.core.base_class.repository import BaseRepository
+from application.core.base_class.repository import BaseAlchemyRepository
 
 from .models import YourModel
 
 session: async_scoped_session = Provide["session"]
 
 
-class YourRepository(BaseRepository):
+class YourAlchemyRepository(BaseAlchemyRepository):
     model: type[YourModel]
 
     def __init__(self, model):
