+++ src/application/domain/domain_structure/service.py
@@ -1,10 +1,10 @@
 from application.core.base_class.service import BaseService
 
-from .repository import YourRepository
+from .repository import YourAlchemyRepository
 
 
 class YourService(BaseService):
-    repository: YourRepository
+    repository: YourAlchemyRepository
 
     def __init__(self, repository):
         super().__init__(repository)
