+++ src/application/domain/home/views.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, Response
 
 from application.core.dependencies import PermissionDependency
 from application.core.fastapi.log_route import LogRoute
-from application.domain.auth.permissions import AllowAll
+from application.core.authority.permissions import AllowAll
 
 home_router = APIRouter(route_class=LogRoute)
