+++ src/application/domain/user/container.py
@@ -1,10 +1,10 @@
 from dependency_injector import containers, providers
 
 from .models import User
-from .repository import UserRepository
+from .repository import UserAlchemyRepository
 from .service import UserService
 
 
 class UserContainer(containers.DeclarativeContainer):
-    user_repository = providers.Factory(UserRepository, model=User)
+    user_repository = providers.Factory(UserAlchemyRepository, model=User)
     user_service = providers.Factory(UserService, repository=user_repository)
