+++ src/application/domain/user/models.py
@@ -4,7 +4,7 @@ from sqlalchemy import BigInteger, Boolean, Column, Enum, Unicode
 from application.core.db import Base
 from application.core.db.mixins import TimestampMixin
 from application.core.fastapi.pydantic_models import ResponseBaseModel
-from application.domain.auth.permissions import Authority
+from application.core.authority.permissions import Authority
 
 
 class User(Base, TimestampMixin):
